refactor: 清理代码并修复潜在问题
- 优化集合操作，使用 addAll 替代循环添加
- 移除冗余的 null 检查和不必要的转换
- 修复正则表达式字符转义问题
- 简化条件逻辑和返回值处理
- 统一工具方法调用，消除重复代码
- 修复 DTO 注解中的冗余 required 属性
- 使用 Math.clamp 和 Math.min 简化数值限制逻辑
- 移除未使用的导入和变量
- 修复可能返回 null 的 Optional 处理
- 使用 Lombok 注解简化 getter/setter
- 修复 switch 表达式以符合 Java 17 语法

+++ src/main/java/com/example/EnterpriseRagCommunity/security/BoardAcl.java
@@ -54,6 +54,6 @@ public class BoardAcl {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         if (auth == null || !auth.isAuthenticated() || "anonymousUser".equals(auth.getPrincipal())) return null;
         String email = auth.getName();
-        return administratorService.findByUsername(email).map(u -> (Long) u.getId()).orElse(null);
+        return administratorService.findByUsername(email).map(u -> u.getId()).orElse(null);
     }
 }
